+++ src/Board/Host/Board.Host.Api/Controllers/FileController.cs
@@ -10,12 +10,13 @@ namespace Board.Host.Api.Controllers;
 /// <summary>
 /// Контроллер для работы с файлами.
 /// </summary>
-/// <response code="500">Произошла внутренняя ошибка.</response>
+
+ 
 [ApiController]
 [Route(template: "files")]
 [Produces("application/json")]
-[AllowAnonymous]
-//[ProducesResponseType(typeof(ErrorDto), StatusCodes.Status500InternalServerError)]
+//[AllowAnonymous]
+
 public class FileController : ControllerBase
 {
     private readonly ILogger<FileController> _logger;
@@ -38,7 +39,14 @@ public class FileController : ControllerBase
     /// <param name="id">Идентификатор файла.</param>
     /// <param name="cancellationToken">Токен отмены.</param>
     /// <returns>Информация о файле.</returns>
+    /// <response code="404"> Нет файла с введенным идентификатором. </response>
+    /// <response code="200"> Информация успешно получена. </response>
+    /// <response code="401"> Пользователь не авторизован. </response>
     [HttpGet("{id}/info")]
+    [Authorize]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     public async Task<IActionResult> GetInfoById(Guid id, CancellationToken cancellationToken)
     {
         var result = await _fileService.GetInfoByIdAsync(id, cancellationToken);
@@ -52,7 +60,12 @@ public class FileController : ControllerBase
     /// <param name="postId"> Идентификатор объявления.</param>
     /// <param name="cancellationToken"> Токен отмены. </param>
     /// <returns> Информация о файле. </returns>
+    /// <response code="404"> Нет файлов или неверный идентификатор. </response>
+    /// <response code="200"> Информация успешно получена. </response>
     [HttpGet("GetAll/{postId:Guid}")]
+    [AllowAnonymous]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
     public async Task<IActionResult> GetAllByPostIdAsync(Guid postId, CancellationToken cancellationToken)
     {
         var result = await _fileService.GetAllByPostIdAsync(postId, cancellationToken);
@@ -67,9 +80,18 @@ public class FileController : ControllerBase
     /// <param name="file">Файл.</param>
     /// <param name="postId"> Идентификатор объявления. </param>
     /// <param name="cancellationToken">Токен отмены.</param>
+    ///// <response code="403"> Нельзя удалить комментарий другого пользователя. </response>
+    /// <response code="404"> Нет файлов или неверный идентификатор. </response>
+    /// <response code="401"> Пользователь не авторизован. </response>
+    /// <response code="201"> Файл успешно загружен. </response>
     [HttpPost]
     [RequestFormLimits(ValueLengthLimit = int.MaxValue, MultipartBodyLengthLimit = long.MaxValue)]
     [DisableRequestSizeLimit]
+    [Authorize]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status201Created)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+    //[ProducesResponseType(StatusCodes.Status403Forbidden)]
     public async Task<IActionResult> Upload(IFormFile file, Guid postId,  CancellationToken cancellationToken)
     {
         var bytes = await GetBytesAsync(file, cancellationToken);
@@ -81,6 +103,9 @@ public class FileController : ControllerBase
             
         };
         var result = await _fileService.UploadAsync(fileDto, postId, cancellationToken);
+
+        if(result == Guid.Parse("00000000-0000-0000-0000-000000000000")) return NotFound();
+
         return StatusCode((int)HttpStatusCode.Created, result);
     }
 
@@ -90,16 +115,19 @@ public class FileController : ControllerBase
     /// <param name="id">Идентификатор файла.</param>
     /// <param name="cancellationToken">Токен отмены.</param>
     /// <returns>Файл в виде потока.</returns>
+    /// <response code="404"> Нет файла с указанным идентификатором. </response>
+    /// <response code="200"> Файл готов к скачиванию. </response>
     [HttpGet("{id}")]
-    //[ProducesResponseType(typeof(ErrorDto), StatusCodes.Status404NotFound)]
+    [AllowAnonymous]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status200OK)]
     public async Task<IActionResult> Download(Guid id, CancellationToken cancellationToken)
     {
         var result = await _fileService.DownloadAsync(id, cancellationToken);
 
         if (result == null)
-        {
             return NotFound();
-        }
+        
 
         Response.ContentLength = result.Content.Length;
         return File(result.Content, result.ContentType, result.Name, true);
@@ -111,12 +139,32 @@ public class FileController : ControllerBase
     /// </summary>
     /// <param name="id">Идентификатор файла.</param>
     /// <param name="cancellationToken">Токен отмены.</param>
+    /// <response code="204"> Файл успешно удалено. </response>
+    /// <response code="404"> Нет файла с введенным идентификатором. </response>
+    /// <response code="403"> Нельзя удалить файл другого пользователя. </response>
+    /// <response code="401"> Пользователь не авторизован. </response>
     [HttpDelete("{id}")]
+    [Authorize]
+    [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(StatusCodes.Status403Forbidden)]
+    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     public async Task<IActionResult> Delete(Guid id, CancellationToken cancellationToken)
     {
+        var existingfile = await _fileService.GetInfoByIdAsync(id, cancellationToken);
+        if (existingfile == null) 
+            return NotFound();
+
+        if(existingfile.AccountId != _fileService.GetCurrentUserId() 
+            || _fileService.GetCurrentUserName() != "Admin")
+            return StatusCode((int)HttpStatusCode.Forbidden);
+
         await _fileService.DeleteAsync(id, cancellationToken);
+
         return NoContent();
     }
+
+
 
     private static async Task<byte[]> GetBytesAsync(IFormFile file, CancellationToken cancellationToken)
     {
